requirements.txt
Cython==0.21.2
kivy
xbmc-json
pyyaml